granule_ingester/conda-requirements.txt
# Licensed to the Apache Software Foundation (ASF) under one or more
# contributor license agreements.  See the NOTICE file distributed with
# this work for additional information regarding copyright ownership.
# The ASF licenses this file to You under the Apache License, Version 2.0
# (the "License"); you may not use this file except in compliance with
# the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

numpy>=1.15.4
scipy
netcdf4>=1.5.3
pandas>=1.0.4
pytz>=2019.3
xarray
pyyaml>=5.3.1
aiohttp==3.6.2
tenacity
requests>=2.27.1
rioxarray
urllib3==1.26.2

